requirements.txt
aiogram==3.4.1
pyrogram==2.0.106
tgcrypto
python-dotenv
requests
